requirements.txt
numpy==1.26.4
scipy==1.13.0
matplotlib==3.9.0
streamlit==1.34.0
opencv-python==4.9.0.80
Pillow==10.3.0
